(PC-22071)[API] feat: split product and events api code
