+++ api/src/pcapi/routes/public/individual_offers/v1/__init__.py
@@ -3,4 +3,5 @@ from flask import Flask
 
 def install_routes(app: Flask) -> None:
     # pylint: disable=unused-import
-    from . import endpoints
+    from . import events
+    from . import products
